refactor: reorganize editor component and associated files
Reorganize the editor component and its related files using a co-located
project structure, placing them within a dedicated 'editor' folder.
