+++ lib/types/bases/misc.ts
@@ -1,18 +1,7 @@
 import { PATH_APP, PATH_HOME } from '@constAssertions/data';
 import { RefObject } from 'react';
-import { Descendant } from 'slate';
-import { ReactEditor } from 'slate-react';
 
-export type CollectTypesMisc = TypesEditor & TypesRouter & TypesWindow & TypesRefs;
-
-export interface TypesEditor {
-  titleName: string;
-  isAutoFocus: boolean;
-  placeholder: string;
-  initialValue: Descendant[];
-  changeHandler: (value: Descendant[]) => void;
-  editor: ReactEditor;
-}
+export type CollectTypesMisc = TypesRouter & TypesWindow & TypesRefs;
 
 export interface TypesRouter {
   path: string;
